move legacy projects to a new folder
