+++ Legacy/2016-12-12/LambdaCsharpToUppercase/global.json
@@ -1,0 +1,6 @@
+{
+  "projects": [ "src", "test" ],
+  "sdk": {
+    "version": "1.0.0-preview2-003131"
+  }
+}
