+++ Legacy/2016-12-12/LambdaCsharpToUppercase/src/LambdaCsharpToUppercase/Function.cs
@@ -1,0 +1,25 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+using Amazon.Lambda.Core;
+using Amazon.Lambda.Serialization;
+
+// Assembly attribute to enable the Lambda function's JSON input to be converted into a .NET class.
+[assembly: LambdaSerializerAttribute(typeof(Amazon.Lambda.Serialization.Json.JsonSerializer))]
+
+namespace LambdaCsharpToUppercase {
+    public class Function {
+
+        /// <summary>
+        /// A simple function that takes a string and does a ToUpper
+        /// </summary>
+        /// <param name="input"></param>
+        /// <param name="context"></param>
+        /// <returns></returns>
+        public string FunctionHandler(string input, ILambdaContext context) {
+            return input?.ToUpper();
+        }
+    }
+}
