+++ Legacy/2016-12-12/LambdaCsharpToUppercase/src/LambdaCsharpToUppercase/aws-lambda-tools-defaults.json
