+++ Legacy/2016-12-12/LambdaCsharpToUppercase/test/LambdaCsharpToUppercase.Tests/FunctionTest.cs
@@ -1,0 +1,28 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+using Xunit;
+using Amazon.Lambda.Core;
+using Amazon.Lambda.TestUtilities;
+
+using LambdaCsharpToUppercase;
+
+namespace LambdaCsharpToUppercase.Tests
+{
+    public class FunctionTest
+    {
+        [Fact]
+        public void TestToUpperFunction()
+        {
+
+            // Invoke the lambda function and confirm the string was upper cased.
+            var function = new Function();
+            var context = new TestLambdaContext();
+            var upperCase = function.FunctionHandler("hello world", context);
+
+            Assert.Equal("HELLO WORLD", upperCase);
+        }
+    }
+}
