+++ Legacy/2016-12-12/LambdaCsharpToUppercase/test/LambdaCsharpToUppercase.Tests/project.json
@@ -1,0 +1,33 @@
+{
+  "version": "1.0.0-*",
+  "buildOptions": {
+  },
+
+  "dependencies": {
+    "Microsoft.NETCore.App": {
+      "type": "platform",
+      "version": "1.0.0"
+    },
+
+    "Amazon.Lambda.Core": "1.0.0*",
+    "Amazon.Lambda.TestUtilities": "1.0.0*",
+ 
+
+
+    "LambdaCsharpToUppercase": "1.0.0",
+
+    "xunit": "2.1.0-*",
+    "dotnet-test-xunit": "2.2.0-*"
+  },
+
+
+
+
+  "testRunner": "xunit",
+
+  "frameworks": {
+    "netcoreapp1.0": {
+      "imports": "dnxcore50"
+    }
+  }
+}
