+++ Legacy/2016-12-12/LambdaJavascriptToUppercase/index.js
