+++ Legacy/2016-12-12/LambdaPythonToUppercase/lambda_function.py
