+++ Legacy/ReadMe.md
@@ -1,0 +1,3 @@
+# Legacy Projects
+
+This folder contains the original Lambda functions to compare performance between Javascript (), Python, and C# (). These projects are only kept for reference and should not be used.
